Personality field + YET ANOTHER UI/CSS rework
Adding the extra table for the personality field confirmed a suspected
issue with legend-as-a-footer: it breaks if the fieldset overflows the
viewport.

One way to fix this could be to have a sticky-like legend, that sits
a the bottom when the user scrolls down, until they move on to the next
fieldset. It is probably doable, but a/ would be pretty hard, since
vanilla "sticky" in CSS wouldn't cut it and b/ I don't like sticky
elements in the first place.

So, instead, I went for a redesign, with the legend taking place of the
big red border. This required fun stuff with text orientation, and some
refactoring (around the --gap custom property mostly) to ensure that
the legend aligns well with the fields labels. A bit of work, but it's
rather pretty now, I think.