Better splitting of the selected option
It doesn't look like much, but this requirement indicated that there
is more logic to our selection than "copy what's in the table". Sometimes
the table entry must by truncated, and the rule to distinguish what's
to be truncated from what's to be kept doesn't look that reliable: it can
be the end of a sentence, it can be the start of a pair of parenthesis…

Something more solid should eventually be found. Either the optional
part is reliably indicated, or the value to be copied is declared
distinctly from the value to display in the table.