More UI/CSS tweaks
But NOW we're REALLY talking!

Since the tables take so much room, moving them to the right only makes
sense. But this implies to rewrite the CSS layout so that everything
can be properly aligned, especially the tables with the input fields.

I would have loved to use a subgrid so that there would be no need to
play with the top margin (and have magic custom properties to like
`--label-height`), but subgrids don't play well with auto-flow, and I
prefer this old-school hack to hardcoding the number of rows in each
fieldset.

Bonus : container queries are cool!