Business logic for relationships
That was interesting. I wanted to go the "high level" route, using
`serialize` instead of the Attributes API, and limiting the amount of
custom types (i.e. the collection itself is a dumb Hash, not something
more business-specific but still hash-like).

Initially, I also defined the DB columns as non-nullabe, with a default
value (by force of habit). But it turns out that `serialize`assumes that
the column is nullable, and actually casts the value to NULL if it
is the same as the default value (see https://github.com/rails/rails/issues/46351)

Besides, having a default value at the DB level doesn't do much, since
AR always passes a value (including sometimes NULL, as explained above).

Another interesting thing is the metaprogramming of the coder module. I
tried with `define_singleton_method` and `instance_eval`, and settled on
the second because I found it more legible. (There may also be performance
benefits, but who care?)

Overall, the code still smells a bit, and I'm starting to wonder if this
"high level/low complexity" approach is, indeed, less complex than having
a custom Type for the Attributes API…