Guess what? EVEN MORE HTML/CSS/UI rework!
With the "secondary fields", we've reached the limits of what our
UI implementation could do – now that we have a better understanding
of how the app should look, feel & behave, we can rethink the
implementation.

The main thing will be to remove the options table from the flow, so
that they can be positionned relative to their associated input
without blocking a lot of vertical space _below_ said input.
To do that, we'll need to stop wrapping the input and its label
by themselves; therefore, the _borders_ and the _indicators_ around the
input cannot be generated based on this wrapper (or the label).

Long story short: the input now needs to be surrounded by two (!) divs,
which is a heartbreaking compromise to do. To make it a bit easier on
the eyes, though, I've added a custom form builder that does the
double wrapping automagically.