These UI redesign/rebuilding never stop…
SO…

With the "subfields" being introduced in the Friends step, the whole
design had to be revised. Even for a similar look & feel, the HTML/CSS
code has been rewritten, so that the subfields will be possible, later
on.

Basically, a "field" component is now the label + the input (with its
border and indicators!) + an optional table (and a description will
be easy to add later). A fieldset is a series of fields, on a grid
whose left column is actually the legend/border.

Fun facts discovered:

-  <legend> can't normally be placed on a grid (there
are not "grid items", unless there display is changed to something which
allows it, such as `grid` or `contents`.
- <caption> cannot contain <button>, if you want to be semantically
correct.

The good thing is that the HTML and the CSS are simpler, now.

STILL TO BE DONE:

- When on a single column, the tables are too close to the input above.
Applying a grid row would also make the label to far away; adding a margin
below the input messes up the buttons on the carousel variant; adding a
top margin to the table isn't right when on 2 columns…