Space matters
I'm really proud of this one. I realized (while working on legend in
several words, such as "tragic love affair") that the spacing between
the words was too small. But, once properly adjusted, it was too big
**specifically between the legend and the ending value in parenthesis**.

So I resorted to typographic nerdery, and used a special whitespace
character, the thin space (see https://qwerty.dev/whitespace/ for
example(s)).