Added role_lifepath
OK, this one is REALLY interesting. And possibly stupid in its current form.

I want to keep all the lifepath-specific stuff in a single attribute,
serialized to a single column. So a lot like for the relationships,
except that it's a single object (not a collection), and that the class
of the object changes with each instance, and cannot be known when the
app loads. Which means that the attribute coder must be somehow universal
(i.e not specific to the serialized object, like the coders for the
relationships are). Which also means that the name of the deserialized
class must be part of the serialized data (like how YAML serialization
works.)

Addding the extra info when serializing isn't too difficult. I went with
overloading as_json instead of doing the adding at the coder lever (in
#dump), or in the object at a higher level (in #to_hash for example),
because this type info is only relevant when converting to JSON.

(OK, you could argue that it's only relevant when converting to JSON
_in order to dump to the DB_, so should be in the .dump method, but
I'm willing to be that it'd be needed when consumming JSON for other
use cases.)

Things are a bit more funky when deserializing, since the role/type/class
now has to be _extracted_ from the JSON before instanciating a new object
(of the right role/type/class). This is my first use case for pattern
matching, as in "hash destructuring", \o/

Another fun thing: because one of the "attributes" of the Fixer class
is a boolean (`work_alone`), and I wanted all the confort of ActiveModel's
attributes (the predicate methods, casting from and to string such as
`"fasle"`, etc.), I included `ActiveModel::Attributes`. This introduced
intersting side-effects, like the serialization adding a `"attributes"`
property to the JSON, because, in the (initial) absence of `to_hash` or
`as_json`, JSON.dump called Object.as_json, which ActiveSupport monkey-
patches to call `instance_variables`, which fetches the, well, instance
variables, among then the `@attributes` added by `AM::Attributes`. Nice
spelunking session, debugging that!