Removed the lockable behavior from options-table
This is getting complex, but quite interesting. The OptionsTable Stimulus
controller was able to do 2 things: pick a value AND be disabled. However,
the disabling in now handled by a different Stimulus controller, which
means that an lockabled options table means to have 2 controller attached:
OptionsTable and Lockable. Which is fine, execpt that the table behavior
must also change when the table is locked (picking an option must not be
possible).

So, the OptionsTable keeps a "picking-disabling" action, but it is *not*
called directly by any DOM element. Instead, the Lockable controller
sends an event, which the OptionsTable listens to.

Good stuff.