Nomad lifepath
This is the tricky one, because picking an option from one field needs
to swap the options for another. Which means that the whole "options
table" field needs to be slightly altered to add the JS controller that
handles the swapping. I've decided to go with duplication this time,
instead of add extra logic to the partial.